admin feature properly implemented in department controller

+++ server/src/main/java/com/cagri/spring/jwt/mongodb/controllers/DepartmentController.java
@@ -1,10 +1,15 @@
 package com.cagri.spring.jwt.mongodb.controllers;
-
+import com.cagri.spring.jwt.mongodb.repository.UserRepository;
+import com.cagri.spring.jwt.mongodb.models.User;
 import com.cagri.spring.jwt.mongodb.models.Department;
 import com.cagri.spring.jwt.mongodb.security.services.DepartmentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -15,6 +20,8 @@ public class DepartmentController {
     @Autowired
     private DepartmentService departmentService;
 
+    @Autowired
+    private UserRepository userRepository;
     // Create department entry
     @PostMapping
     public ResponseEntity<Department> createDepartment(@RequestBody Department department) {
@@ -46,7 +53,32 @@ public class DepartmentController {
     // Delete department entry
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteDepartment(@PathVariable String id) {
+        String currentUsername = getAuthenticatedUsername();
+        String currentUserRole = getRoleByUsername(currentUsername);
+        if (!"admin".equalsIgnoreCase(currentUserRole)) {
+            return ResponseEntity.status(403).build(); // Forbidden if the user is not an admin
+        }
+
         departmentService.deleteDepartment(id);
         return ResponseEntity.noContent().build();
+    }
+
+    // Helper method to get the role of the currently authenticated user
+    private String getAuthenticatedUserRole() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getAuthorities().stream()
+                .findFirst() // Assuming a single role per user
+                .map(GrantedAuthority::getAuthority)
+                .orElse(null);
+    }
+    private String getAuthenticatedUsername() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getName();
+    }
+    private String getRoleByUsername(String username) {
+        User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new RuntimeException("Error: User not found!"));
+
+        return user.getRole();
     }
 }
